# web/openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestRunRequest": {
        "additionalProperties": false,
        "description": "创建单策略、单品种回测所需的可信输入。",
        "properties": {
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "fee_bps": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,11}0*$)\\d{0,4}\\.\\d{0,6}0*$)",
                "type": "string"
              }
            ],
            "title": "Fee Bps"
          },
          "initial_capital": {
            "anyOf": [
              {
                "maximum": 1000000000000.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,30}|(?=[\\d.]{1,31}0*$)\\d*\\.\\d*0*$)",
                "type": "string"
              }
            ],
            "title": "Initial Capital"
          },
          "leverage": {
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Leverage",
            "type": "integer"
          },
          "max_holding_bars": {
            "maximum": 50000.0,
            "minimum": 0.0,
            "title": "Max Holding Bars",
            "type": "integer"
          },
          "params": {
            "additionalProperties": {
              "type": "number"
            },
            "maxProperties": 64,
            "title": "Params",
            "type": "object"
          },
          "position_size_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.01,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,11}0*$)\\d{0,4}\\.\\d{0,6}0*$)",
                "type": "string"
              }
            ],
            "title": "Position Size Pct"
          },
          "slippage_bps": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,11}0*$)\\d{0,4}\\.\\d{0,6}0*$)",
                "type": "string"
              }
            ],
            "title": "Slippage Bps"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "stop_loss_pct": {
            "anyOf": [
              {
                "maximum": 99.9,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,11}0*$)\\d{0,4}\\.\\d{0,6}0*$)",
                "type": "string"
              }
            ],
            "title": "Stop Loss Pct"
          },
          "strategy_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]*$",
            "title": "Strategy Id",
            "type": "string"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 2,
            "pattern": "^[A-Z0-9][A-Z0-9._:/-]*$",
            "title": "Symbol",
            "type": "string"
          },
          "take_profit_pct": {
            "anyOf": [
              {
                "maximum": 99.9,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,11}0*$)\\d{0,4}\\.\\d{0,6}0*$)",
                "type": "string"
              }
            ],
            "title": "Take Profit Pct"
          },
          "timeframe": {
            "enum": [
              "1m",
              "3m",
              "5m",
              "15m",
              "30m",
              "1h",
              "2h",
              "4h",
              "6h",
              "8h",
              "12h",
              "1d",
              "3d",
              "1w"
            ],
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "strategy_id",
          "symbol",
          "timeframe",
          "start_date",
          "end_date",
          "initial_capital",
          "position_size_pct",
          "leverage",
          "fee_bps",
          "slippage_bps",
          "stop_loss_pct",
          "take_profit_pct",
          "max_holding_bars"
        ],
        "title": "BacktestRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "QuantDesk V2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/backtests": {
      "get": {
        "operationId": "list_backtests_api_v2_backtests_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 12,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "title": "Response List Backtests Api V2 Backtests Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Backtests"
      },
      "post": {
        "operationId": "create_backtest_api_v2_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Backtest Api V2 Backtests Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Backtest"
      }
    },
...
    "/api/v2/backtests/{run_id}": {
      "get": {
        "operationId": "get_backtest_api_v2_backtests__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Backtest Api V2 Backtests  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Backtest"
      }
    },
...
    "/api/v2/monitor/breadth": {
      "get": {
        "operationId": "monitor_breadth_api_v2_monitor_breadth_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Breadth Api V2 Monitor Breadth Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Breadth"
      }
    },
    "/api/v2/monitor/intelligence": {
      "get": {
        "operationId": "monitor_intelligence_api_v2_monitor_intelligence_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Intelligence Api V2 Monitor Intelligence Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Intelligence"
      }
    },
    "/api/v2/monitor/klines": {
      "get": {
        "operationId": "monitor_klines_api_v2_monitor_klines_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tf",
            "required": false,
            "schema": {
              "default": "1h",
              "pattern": "^(15m|1h|4h)$",
              "title": "Tf",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 120,
              "maximum": 300,
              "minimum": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Monitor Klines Api V2 Monitor Klines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Klines"
      }
    },
    "/api/v2/monitor/news": {
      "get": {
        "operationId": "monitor_news_api_v2_monitor_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 60,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Monitor News Api V2 Monitor News Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor News"
      }
    },
...
    "/api/v2/monitor/overview": {
      "get": {
        "operationId": "monitor_overview_api_v2_monitor_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Overview Api V2 Monitor Overview Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Overview"
      }
    },
...
    "/api/v2/monitor/report": {
      "get": {
        "operationId": "monitor_report_api_v2_monitor_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Report Api V2 Monitor Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Report"
      }
    },
    "/api/v2/monitor/score": {
      "get": {
        "operationId": "monitor_score_api_v2_monitor_score_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Monitor Score Api V2 Monitor Score Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitor Score"
      }
    },
...
  }
}